+++ crates/shared/src/dto/user.rs
@@ -12,7 +12,7 @@ pub struct User {
     pub id: String,
     pub username: String,
     pub created_at: DateTime<Utc>,
-    pub elo: i32,
+    pub elo: f64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -20,7 +20,7 @@ pub struct PrivateUser {
     pub id: String,
     pub username: String,
     pub created_at: DateTime<Utc>,
-    pub elo: i32,
+    pub elo: f64,
     pub competing_robot_id: Option<String>,
 }
